+++ libraries/chain/wasm_price.hpp
@@ -94,7 +94,7 @@ namespace contento { namespace chain { namespace wasm_price {
 
     WASM_PRICE_FN_DECLS(
                          (get_contract_balance,      (int64_t) )
-                         (transfer,                  (void*, int64_t, int64_t))
+                         (transfer,                  (void*, int64_t, int))
                          (get_value,                 (int64_t))
                          (pay_prohibited,            (void*))
                          );
